+++ src/templates/Home/lazy-component.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+const LazyComponent = () => {
+  return (
+    <div>
+      <p>LazyComponent</p>
+    </div>
+  );
+};
+
+export default LazyComponent;
